part1-database-etl/requirements.txt
pandas>=1.5.0
mysql-connector-python>=8.0.30
numpy>=1.21.0
python-dateutil>=2.8.2
jupyter- 7.4.5
python= 3.10 (via Anaconda Distribution)